\name{bio.tagging-package}
\alias{bio.tagging-package}
\alias{bio.tagging}
\docType{package}
\title{
\packageTitle{bio.tagging}
}
\description{
\packageDescription{bio.tagging}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bio.tagging}
\packageIndices{bio.tagging}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bio.tagging}

Maintainer: \packageMaintainer{bio.tagging}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
